+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Inspired by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Note: this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - Unreleased
+### Added
+- Added a rake task definition that can be optionally included into a non-Rails project to generate
+  schema migrations.
+
 ## [1.1.0] - 2022-07-22
 ### Changed
 - Fixed a bug where `DeclareSchema::Model::HabtmModelShim` `indexes` and `integer limits` were not being generated properly. Use `limit 8` for ids and primary composite key for habtm model.
@@ -229,6 +234,7 @@ using the appropriate Rails configuration attributes.
 ### Added
 - Initial version from https://github.com/Invoca/hobo_fields v4.1.0.
 
+[1.2.0]: https://github.com/Invoca/declare_schema/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/Invoca/declare_schema/compare/v1.0.2...v1.1.0
 [1.0.2]: https://github.com/Invoca/declare_schema/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/Invoca/declare_schema/compare/v1.0.0...v1.0.1